Don't filter starlark flag resolution based on --build_tests_only and friends
Previously when --build_tests_only was passed alongside any starlark flag a crash would occur because the starlark flag is not a test and therefore would be filtered from the targets available at the top level.

With this change no flags that affect top-level target availability should impact the loading of starlark flags.

RELNOTES: None
PiperOrigin-RevId: 281397052